More multichamber work
