+++ .gitignore
@@ -114,8 +114,6 @@ _UpgradeReport_Files/
 Backup*/
 UpgradeLog*.XML
 
-
-
 ############
 ## Windows
 ############
@@ -125,7 +123,6 @@ Thumbs.db
 
 # Folder config file
 Desktop.ini
-
 
 #############
 ## Python
@@ -170,3 +167,5 @@ userSettings.json
 do_not_run_brewpi
 config_user.php
 .idea/*
+lcd.php
+
